@@ -1,46 +1,46 @@
 import * as React from 'react';
-import { Theme, withStyles } from '@material-ui/core/styles';
-import Link, { LinkProps } from '@material-ui/core/Link';
-import { SFGrey } from '../../SFColors/SFColors';
-import { CSSProperties } from '@material-ui/core/styles/withStyles';
+import { styled, Link, LinkProps } from '@mui/material';
+import {
+  SFBlueMainDark,
+  SFBlueMainLight,
+  SFGrey
+} from '../../SFColors/SFColors';
+import { SFTheme } from '../../SFTheme/SFTheme';
 
-type SFSize = 'small' | 'medium';
-const StyledLink = withStyles((theme: Theme) => ({
-  root: {
-    '&.MuiTypography-colorInherit': {
-      color: `${theme.palette.type === 'light' ? SFGrey[900] : SFGrey[50]}`
-    },
-    '&.MuiLink-root': {
-      cursor: 'pointer',
-      fontWeight: 500
-    }
+function getColor(sfColor: SFLinkColor, theme: SFTheme): string {
+  if (theme.palette.mode === 'light') {
+    if (sfColor === 'primary') return SFBlueMainLight;
+    else return SFGrey[900];
+  } else {
+    if (sfColor === 'primary') return SFBlueMainDark;
+    else return SFGrey[50];
   }
-}))(Link);
-
-const getSizeStyle = (size?: SFSize): CSSProperties => {
-  switch (size) {
-    case 'small':
-      return {
-        fontSize: '14px',
-        lineHeight: '20px'
-      };
-    default:
-      return {
-        fontSize: '16px',
-        lineHeight: '24px'
-      };
-  }
-};
-
-export interface SFLinkProps extends LinkProps {
-  sfSize: SFSize;
 }
 
-export const SFLink = ({
+export type SFLinkSize = 'small' | 'medium';
+export type SFLinkColor = 'default' | 'primary';
+
+export interface SFLinkProps extends LinkProps {
+  sfSize?: SFLinkSize;
+  sfColor?: SFLinkColor;
+}
+
+const SFLinkBase = ({
   sfSize = 'medium',
+  sfColor = 'default',
   color,
   ...props
 }: SFLinkProps): React.ReactElement<SFLinkProps> => {
-  const sfColor = color !== 'primary' ? 'inherit' : 'primary';
-  return <StyledLink {...props} style={getSizeStyle(sfSize)} color={sfColor} />;
+  return <Link {...props} />;
 };
+
+export const SFLink = styled(SFLinkBase)(
+  ({ sfSize, sfColor = 'default', theme }) => ({
+    cursor: 'pointer',
+    fontWeight: 500,
+    textDecorationColor: 'unset',
+    color: getColor(sfColor, theme),
+    fontSize: sfSize === 'small' ? '14px' : '16px',
+    lineHeight: sfSize === 'small' ? '20px' : '24px'
+  })
+);
