@@ -1,4 +1,4 @@
-import { createTheme } from '@material-ui/core/styles';
+import { createTheme } from '@mui/material/styles';
 import { CommonThemeOptions } from './CommonThemeOptions';
 
 import {
@@ -12,7 +12,7 @@ import {
 } from '../SFColors/SFColors';
 
 CommonThemeOptions.palette = {
-  type: 'dark',
+  mode: 'dark',
   text: {
     primary: SFTextWhite
   },
