@@ -2,7 +2,7 @@ import {
   Theme,
   ThemeOptions,
   ThemeProvider as SFThemeProvider
-} from '@material-ui/core/styles';
+} from '@mui/material/styles';
 
 import { NightTheme } from './NightTheme';
 import { DayTheme } from './DayTheme';
