@@ -1,23 +1,21 @@
 import React from 'react';
-// also exported from '@storybook/react' if you can deal with breaking changes in 6.1
-import { Story, Meta } from '@storybook/react/types-6-0';
+import { Story, ComponentMeta } from '@storybook/react';
 import { SFLink, SFLinkProps } from './SFLink';
 
 export default {
   title: 'Components/SFLink',
   component: SFLink,
-  parameters: { controls: { sort: 'alpha' } },
+  parameters: {
+    controls: { sort: 'alpha', include: ['sfColor', 'sfSize', 'text'] }
+  },
   args: {
     text: 'I am a link example',
-    color: 'inherited',
+    sfColor: 'default',
     sfSize: 'medium'
   },
   argTypes: {
     onClick: {
-      action: 'onClick',
-      table: {
-        disable: true
-      }
+      action: 'onClick'
     },
     text: {
       description: 'The text to display.',
@@ -25,69 +23,19 @@ export default {
         type: 'text'
       }
     },
-    color: {
+    sfColor: {
       description: 'The color of the link.',
-      options: ['inherited', 'primary'],
+      options: ['default', 'primary'],
       control: {
         type: 'radio'
       }
     },
     sfSize: {
       description: 'The size of the link.'
-    },
-    variant: {
-      table: {
-        disable: true
-      }
-    },
-    noWrap: {
-      table: {
-        disable: true
-      }
-    },
-    gutterBottom: {
-      table: {
-        disable: true
-      }
-    },
-    paragraph: {
-      table: {
-        disable: true
-      }
-    },
-    align: {
-      table: {
-        disable: true
-      }
-    },
-    display: {
-      table: {
-        disable: true
-      }
-    },
-    underline: {
-      table: {
-        disable: true
-      }
-    },
-    ref: {
-      table: {
-        disable: true
-      }
-    },
-    variantMapping: {
-      table: {
-        disable: true
-      }
-    },
-    TypographyClasses: {
-      table: {
-        disable: true
-      }
     }
   }
-} as Meta;
+} as ComponentMeta<typeof SFLink>;
 
 export const Default: Story = ({ text, ...args }) => {
-  return <SFLink {...(args as SFLinkProps)}>{text}</SFLink>;
+  return <SFLink {...args}>{text}</SFLink>;
 };
